move calendarjs to admin dir
