+++ resources/index.js
@@ -1,7 +1,7 @@
 import 'htmx.org'
 import './index.scss'
 import { initDatepicker } from "./flatpickr.js"
-import { initCalendar } from './calendar'
+import { initCalendar } from './admin/calendar'
 import {
     initConfirmable,
     initFileInput,
